backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic[email]>=2.5.0
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
passlib[bcrypt]>=1.7.4
requests>=2.31.0
email-validator>=2.1.0
openai>=1.0.0
python-multipart>=0.0.6